refactor(discover): actualizar tipos y rutas de importación en componentes de descubrimiento
- Se modificaron los tipos de contenido multimedia en varios componentes, cambiando de 'series' y 'movies' a 'serie' y 'pelicula' para una mejor consistencia en el código.
- Se actualizaron las rutas de importación para `MediaType` y `TrendingSerie`, mejorando la organización y claridad del código.
- Estos cambios optimizan la estructura y la legibilidad en la sección de descubrimiento.
